+++ RELEASES.md
@@ -6,7 +6,7 @@ Released on TBC.
 - Removed redundant `realized_points` concept from `Position`
 
 ### Enhancements
-- Added `Cache.archive_position`
+- Added `Cache.snapshot_position`
 - Execution engine now takes snapshots of closed netted positions
 - Performance statistics now based on total positions and snapshots
 - Added Binance Spot/Margin external order handling